Extract XLSX common code from conference scheduling example
