+++ optaplanner-examples/src/main/java/org/optaplanner/examples/common/persistence/generator/StringDataGenerator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.examples.common.persistence;
+package org.optaplanner.examples.common.persistence.generator;
 
 import java.util.ArrayList;
 import java.util.List;
